refactor: resize이벤트 기반 높이를 구하는 로직을 재사용하기 위한 커스텀 hook 분리
